feat: add admin-panel route
