+++ stores/admin/index.ts
@@ -1,0 +1,5 @@
+export const useAdminStore = defineStore('admin', () => {
+  const isAdmin = ref(false)
+
+  return { isAdmin }
+})
