--- stores/user/index.ts
@@ -1,5 +1,0 @@
-export const useUserStore = defineStore('user', () => {
-  const isAdmin = ref(false)
-
-  return { isAdmin }
-})
